starklings-backend/requirements.txt
bcrypt==3.2.2
Flask==2.1.3
flask-cors==3.0.10
aioflask==0.4.0
Flask-SQLAlchemy==2.5.1
mysqlclient==2.1.1
pathspec==0.9.0
PyJWT==2.4.0
PyMySQL==1.0.2
python-dotenv==0.20.0
requests==2.28.1
rich==12.4.4
SQLAlchemy==1.4.39
typing_extensions==4.3.0
urllib3==1.26.11
pytest==7.1.2
